test(postgres): keep scoped sync-state tests untagged
The sync-state migration checks do not require a live PostgreSQL server, so keeping them behind the pgtest tag leaves the default test suite without coverage for the copy-then-clear safety path.

Move the SQLite-only tests and helper into the untagged sync unit test file, while keeping the PostgreSQL integration coverage in the pgtest file. Also align single-target pg push output with pg status so explicitly selected targets are visible in command output.
